RandomizeWithinISA

This function is used in close connection with InitializeISARandomizer and if a new version of this fuction is developed a new version InitializeISARandomizer with the same class is required. RandomizeWithInISA should return a list with at least the following items nTrt - The treatment within the ISA that the patient should receive cISARand - the same is cISARand but if it is updates this allows the caller to get the update.

RandomizeWithinISA(cISARand, dCurrentTime)

# S3 method for default
RandomizeWithinISA(cISARand, dCurrentTime)

# S3 method for EqualRandomizer
RandomizeWithinISA(cISARand, dCurrentTime)

# S3 method for POCRandomizer
RandomizeWithinISA(cISARand, dCurrentTime)

# S3 method for DelayedStartRandomizer
RandomizeWithinISA(cISARand, dCurrentTime)

Methods (by class)

default: Because several options are provided and there in no well defined default an stop error occurs if you call the default method.

EqualRandomizer: Patients are randomized equally/fairly amount the treatments in the ISA.

POCRandomizer: This randomzier will randomize the first patients to a select set of arms or doses in the ISA (eg a POC phase). The remaining arms or doses are opened based on the number of intitial patients.

DelayedStartRandomizer: This version will randomize the first patients to a select set of arms or doses in the ISA (eg a POC phase) The remaining arms or doses are opened based on the provided times to open.

See also